move `AvailableProjectPaths` to `project` package
