+++ internal/runbits/runbits.go
@@ -1,0 +1,2 @@
+// Package runbits comprises logic that is shared between controllers, ie., code that prints
+package runbits
